src/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
cssutils==2.2.0
currency-symbols==2.0.0
et-xmlfile==1.1.0
fonttools==4.22.1
idna==2.10
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
urllib3==1.26.4
webcolors==1.11.1
xlrd==2.0.1
